# docs/swagger.yaml
basePath: /
definitions:
...
  page.CreateInput:
    properties:
      content:
        type: string
      is_published:
        type: boolean
      slug:
        maxLength: 255
        minLength: 2
        type: string
      title:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - title
    type: object
  page.Model:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_published:
        type: boolean
      slug:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  page.UpdateInput:
    properties:
      content:
        type: string
      is_published:
        type: boolean
      slug:
        maxLength: 255
        minLength: 2
        type: string
      title:
        maxLength: 255
        minLength: 2
        type: string
    type: object
  payment.CreateInput:
    properties:
      payment_method:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - payment_method
    type: object
  payment.Model:
    properties:
      amount:
        type: number
      course_id:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      paid_at:
        type: string
      payment_method:
        type: string
      status:
        type: string
      transaction_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  payment.UpdateStatusInput:
    properties:
      status:
        enum:
        - completed
        - failed
        - refunded
        type: string
      transaction_id:
        type: string
    required:
    - status
    type: object
...
  user.DeleteInput:
    properties:
      password:
        type: string
    type: object
  user.ErrorResponse:
    properties:
      error:
        example: email is required
        type: string
    type: object
  user.Model:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  user.UpdateInput:
    properties:
      email:
        example: ivan@example.com
        type: string
      first_name:
        example: Ivanov
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        example: Ivanovich
        maxLength: 100
        minLength: 2
        type: string
      name:
        example: Ivan
        maxLength: 100
        minLength: 2
        type: string
      phone:
        example: "9094445566"
        maxLength: 20
        minLength: 10
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: Документация API сервиса it_courses.
  title: API it_courses
  version: "1.0"
paths:
...
  /courses/{id}/payments:
    post:
      consumes:
      - application/json
      description: Создаёт платёж со статусом pending. Только student
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      - description: Данные платежа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/payment.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/payment.Model'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      summary: Создать платёж за курс
      tags:
      - payments
...
  /pages:
    get:
      description: Возвращает список всех статических страниц
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/page.Model'
            type: array
      summary: Список страниц
      tags:
      - pages
    post:
      consumes:
      - application/json
      description: Создаёт новую статическую страницу. Доступно teacher и admin
      parameters:
      - description: Данные страницы
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/page.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/page.Model'
        "400":
          description: Bad Request
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      summary: Создать страницу
      tags:
      - pages
  /pages/{id}:
    delete:
      description: Удаляет статическую страницу. Доступно teacher и admin
      parameters:
      - description: ID страницы
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: string
      summary: Удалить страницу
      tags:
      - pages
    put:
      consumes:
      - application/json
      description: Обновляет статическую страницу. Доступно teacher и admin
      parameters:
      - description: ID страницы
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/page.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/page.Model'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      summary: Обновить страницу
      tags:
      - pages
  /pages/{slug}:
    get:
      description: Возвращает статическую страницу по slug
      parameters:
      - description: Slug страницы
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/page.Model'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Детали страницы
      tags:
      - pages
...
  /payments/{id}:
    get:
      description: Возвращает платёж по ID. Student — только свои, admin — любые
      parameters:
      - description: ID платежа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.Model'
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Детали платежа
      tags:
      - payments
  /payments/{id}/status:
    patch:
      consumes:
      - application/json
      description: Меняет статус платежа. При completed автоматически выдаётся доступ
        к курсу. Только admin
      parameters:
      - description: ID платежа
        in: path
        name: id
        required: true
        type: integer
      - description: Новый статус
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/payment.UpdateStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.Model'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      summary: Обновить статус платежа
      tags:
      - payments
...
  /users:
    get:
      description: Возвращает список всех пользователей. Только admin
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.Model'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.ErrorResponse'
      summary: Список пользователей
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет пользователя. Владелец (с подтверждением пароля) или admin
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Пароль (только для self-delete)
        in: body
        name: input
        schema:
          $ref: '#/definitions/user.DeleteInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/user.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/user.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/user.ErrorResponse'
      summary: Удалить пользователя
      tags:
      - users
    get:
      description: Возвращает данные пользователя по ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.Model'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/user.ErrorResponse'
      summary: Получить пользователя
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет данные пользователя. Владелец или admin
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.Model'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/user.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/user.ErrorResponse'
      summary: Обновить пользователя
      tags:
      - users
swagger: "2.0"